+++ tests/testthat/test-pact_list.R
@@ -2,9 +2,10 @@
 
 ## Tests ----
 test_that("pact_list outputs as expected", {
-  skip_if(Sys.getenv("FIGSHARE_TOKEN") == "")
   skip_on_cran()
   skip_if_offline()
+  skip_if(Sys.getenv("PACTR_INTEGRATION_TESTS") == FALSE)
+  skip_if(Sys.getenv("FIGSHARE_TOKEN") == "")
 
   pact_client <- pact_client_set()