+++ NEWS.md
@@ -1,4 +1,4 @@
-# pactr v0.0.0.9008 (development version)
+# pactr v0.0.0.9009 (development version)
 
 This is a pre-release, development version. In this release:
 
@@ -17,6 +17,8 @@ This is a pre-release, development version. In this release:
 - `pact_table_*()` functions for creating tables from the Pandemic PACT data have been updated to reflect the new data structure.
 
 - `pact_figshare_list()` replaces `pact_list()` to signify that that is a Figshare specific function. `pact_figshare_list()` has additional functionality to retrieve and return all available outputs/assets from the Figshare repository of Pandemic PACT.
+
+- `pact_data_read_figshare()` function for reading data from the Pandemic PACT Figshare repository has replaced the older `pact_read_figshare()` function.
 
 - package tests have been refactored so as to limit API calls and network calls through use of mock tests whenever possible.